+++ bot.py
@@ -1,0 +1,4 @@
+from vkbottle.bot import Message, Bot
+from config import vk_token
+
+bot = Bot(vk_token)
